Updated file system to reflect flowchart plan & removed some old test files.
