+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/LOADCode/Main_/Hardware_/Drivetrain_/MecanumDrivetrainClass.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.LOADCode.Main_.Hardware_Classes;
+package org.firstinspires.ftc.teamcode.LOADCode.Main_.Hardware_.Drivetrain_;
 
 import androidx.annotation.NonNull;
