+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/LOADCode/Testing_/PedroDriveToAprilTag.java
@@ -27,17 +27,15 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.LOADCode.Testing_.TeleOps;
+package org.firstinspires.ftc.teamcode.LOADCode.Testing_;
 
 import com.bylazar.configurables.annotations.IgnoreConfigurable;
 import com.bylazar.telemetry.TelemetryManager;
 import com.pedropathing.follower.Follower;
 import com.pedropathing.geometry.Pose;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import com.qualcomm.robotcore.hardware.AnalogInput;
-import com.qualcomm.robotcore.hardware.CRServo;
+import com.qualcomm.robotcore.util.Range;
 
 import org.firstinspires.ftc.robotcore.external.hardware.camera.BuiltinCameraDirection;
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
@@ -86,65 +84,72 @@ import java.util.concurrent.TimeUnit;
  * Use DESIRED_DISTANCE to set how close you want the robot to get to the target.
  * Speed and Turn sensitivity can be adjusted using the SPEED_GAIN, STRAFE_GAIN and TURN_GAIN constants.
  *
+ * Use Android Studio to Copy this Class, and Paste it into the TeamCode/src/main/java/org/firstinspires/ftc/teamcode folder.
+ * Remove or comment out the @Disabled line to add this OpMode to the Driver Station OpMode list.
+ *
  */
-@Disabled
-@TeleOp(name="Turret April-Tag Tracking", group = "TestTeleOp")
-public class TurretAprilTagTracking extends LinearOpMode
+@TeleOp(name="Follow AprilTag", group = "TestTeleOp")
+public class PedroDriveToAprilTag extends LinearOpMode
 {
+    // Adjust these numbers to suit your robot.
+    final double DESIRED_DISTANCE = 60.0; //  this is how close the camera should get to the target (inches)
+
+    //  Set the GAIN constants to control the relationship between the measured position error, and how much power is
+    //  applied to the drive motors to correct the error.
+    //  Drive = Error * Gain    Make these values smaller for smoother control, or larger for a more aggressive response.
+    final double SPEED_GAIN  =  0.04  ;   //  Forward Speed Control "Gain". e.g. Ramp up to 50% power at a 25 inch error.   (0.50 / 25.0)
+    final double STRAFE_GAIN =  0.04 ;   //  Strafe Speed Control "Gain".  e.g. Ramp up to 37% power at a 25 degree Yaw error.   (0.375 / 25.0)
+    final double TURN_GAIN   =  0.04  ;   //  Turn Control "Gain".  e.g. Ramp up to 25% power at a 25 degree error. (0.25 / 25.0)
+
+    final double MAX_AUTO_SPEED = 1;   //  Clip the approach speed to this max value (adjust for your robot)
+    final double MAX_AUTO_STRAFE= 1;   //  Clip the strafing speed to this max value (adjust for your robot)
+    final double MAX_AUTO_TURN  = 1;   //  Clip the turn speed to this max value (adjust for your robot)
+
     private static final boolean USE_WEBCAM = true;  // Set true to use a webcam, or false for a phone camera
-    private static final int DESIRED_TAG_ID = 24;    // Choose the tag you want to approach or set to -1 for ANY tag.
+    private static final int DESIRED_TAG_ID = -1;     // Choose the tag you want to approach or set to -1 for ANY tag.
     private VisionPortal visionPortal;               // Used to manage the video source.
     private AprilTagProcessor aprilTag;              // Used for managing the AprilTag detection process.
-    public static Follower follower;                 // Used for managing the PedroPathing path follower
-    @IgnoreConfigurable
-    public TelemetryManager telemetryM;              // Used for putting telemetry data on Panels
+    private AprilTagDetection desiredTag = null;     // Used to hold the data for a detected AprilTag
 
-    // Contains the start Pose of our robot. This can be changed or saved from the autonomous period.
-    private final Pose startPose = new Pose(60,96, Math.toRadians(0));
+    private boolean toggle = false;
+    private boolean oldToggleVal = false;
+
+    public static Follower follower;
+    @IgnoreConfigurable
+    static TelemetryManager telemetryM;
+    /** Start Pose of our robot. This can be changed or saved from the autonomous period. */
+    private final Pose startPose = new Pose(0,0, Math.toRadians(0));
 
     @Override public void runOpMode()
     {
-        boolean targetFound;        // Set to true when an AprilTag target is detected
-        double headingError = 0;    // The error in degrees in the turret's angle
+        boolean targetFound;    // Set to true when an AprilTag target is detected
+        double  drive;          // Desired forward power/speed (-1 to +1)
+        double  strafe;         // Desired strafe power/speed (-1 to +1)
+        double  turn;           // Desired turning power/speed (-1 to +1)
 
-        // Used to store the hardware objects for the robot
-        CRServo turret;
-        AnalogInput turretEncoder;
-
-        double turretPos;           // Used to store the current angle of the turret
-        double turretPower;     // Used to store the calculated power to output to the turret servo
-
-        initAprilTag();             // Initialize the Apriltag Detection process
+        // Initialize the Apriltag Detection process
+        initAprilTag();
 
         if (USE_WEBCAM)
             setManualExposure(6, 250);  // Use low exposure time to reduce motion blur
 
-        follower = Constants.createFollower(hardwareMap);   // Initializes the PedroPathing path follower
-        follower.setStartingPose(startPose);                // Sets the initial position of the robot on the field
-        follower.update();                                  // Applies the initialization
+        follower = Constants.createFollower(hardwareMap);
+        follower.setStartingPose(startPose);
+        follower.update();
 
+        // Wait for driver to press start
         telemetry.addData("Camera preview on/off", "3 dots, Camera Stream");
         telemetry.addData(">", "Touch START to start OpMode");
         telemetry.update();
+        waitForStart();
 
-        waitForStart(); // Wait for driver to press start
-
-        follower.startTeleopDrive();    // Activate the manual drive function for the drivetrain
-        follower.update();              // Apply the activation
-
-        // Fetch the actual hardware objects and store them in their respective variables
-        turret = hardwareMap.get(CRServo.class, "turret");
-        turretEncoder = hardwareMap.get(AnalogInput.class,"turretEncoder");
+        follower.startTeleopDrive();
+        follower.update();
 
         while (opModeIsActive())
         {
-            // Used to indicate whether or not a valid AprilTag has been detected
             targetFound = false;
-            // Used to hold the data for a detected AprilTag
-            AprilTagDetection desiredTag = null;
-
-            // Read the Axon servo encoder position and convert it from a voltage to an angle in degrees
-            turretPos = (turretEncoder.getVoltage() / 3.3) * 360;
+            desiredTag  = null;
 
             // Step through the list of detected tags and look for a matching tag
             List<AprilTagDetection> currentDetections = aprilTag.getDetections();
@@ -156,8 +161,6 @@ public class TurretAprilTagTracking extends LinearOpMode
                         // Yes, we want to use this tag.
                         targetFound = true;
                         desiredTag = detection;
-                        // Determine heading error so we can use them to control the robot automatically.
-                        headingError = -desiredTag.ftcPose.bearing;
                         break;  // don't look any further.
                     } else {
                         // This tag is in the library, but we do not want to track it right now.
@@ -171,39 +174,50 @@ public class TurretAprilTagTracking extends LinearOpMode
 
             // Tell the driver what we see, and what to do.
             if (targetFound) {
+                telemetry.addData("\n>","HOLD Left-Bumper to Drive to Target\n");
                 telemetry.addData("Found", "ID %d (%s)", desiredTag.id, desiredTag.metadata.name);
                 telemetry.addData("Range",  "%5.1f inches", desiredTag.ftcPose.range);
                 telemetry.addData("Bearing","%3.0f degrees", desiredTag.ftcPose.bearing);
                 telemetry.addData("Yaw","%3.0f degrees", desiredTag.ftcPose.yaw);
             } else {
-                telemetry.addData("\n>","Drive using joysticks to find a valid target\n");
+                telemetry.addData("\n>","Drive using joysticks to find valid target\n");
             }
 
-            // If B on Gamepad1 is being pressed, AND we have found the desired target, automatically aim the turret at the AprilTag.
-            // Otherwise, stop moving
-            if (gamepad1.b) {
-                if (Math.abs(headingError) >= 0.5) {
-                    turretPower = Math.pow(Math.min(Math.abs(headingError/7.5),1), 2) * Math.signum(headingError) * 0.2;
-                } else {
-                    turretPower = 0;
-                }
+            if (gamepad1.b && !oldToggleVal){
+                toggle = !toggle;
+                oldToggleVal = true;
+            }else if (!gamepad1.b && oldToggleVal){
+                oldToggleVal = false;
+            }
+
+            // If Left Bumper is being pressed, AND we have found the desired target, Drive to target Automatically .
+            if (toggle && targetFound) {
+
+                // Determine heading, range and Yaw (tag image rotation) error so we can use them to control the robot automatically.
+                double  rangeError      = (desiredTag.ftcPose.range - DESIRED_DISTANCE);
+                double  headingError    = desiredTag.ftcPose.bearing;
+                double  yawError        = desiredTag.ftcPose.yaw;
+
+                // Use the speed and turn "gains" to calculate how we want the robot to move.
+                drive  = Range.clip(rangeError * SPEED_GAIN, -MAX_AUTO_SPEED, MAX_AUTO_SPEED);
+                turn   = Range.clip(headingError * TURN_GAIN, -MAX_AUTO_TURN, MAX_AUTO_TURN) ;
+                strafe = Range.clip(-yawError * STRAFE_GAIN, -MAX_AUTO_STRAFE, MAX_AUTO_STRAFE);
+
+                telemetry.addData("Auto","Drive %5.2f, Strafe %5.2f, Turn %5.2f ", drive, strafe, turn);
             } else {
-                turretPower = 0;
+
+                // drive using manual POV Joystick mode.  Slow things down to make the robot more controlable.
+                drive  = -gamepad1.left_stick_y  / 2.0;  // Reduce drive rate to 50%.
+                strafe = -gamepad1.left_stick_x  / 2.0;  // Reduce strafe rate to 50%.
+                turn   = -gamepad1.right_stick_x / 3.0;  // Reduce turn rate to 33%.
+                telemetry.addData("Manual","Drive %5.2f, Strafe %5.2f, Turn %5.2f ", drive, strafe, turn);
             }
-            if(gamepad1.b){
-                turret.setPower(turretPower);
-            }
+            telemetry.update();
 
             // Apply desired axes motions to the drivetrain.
-            follower.setTeleOpDrive(-gamepad1.left_stick_y, -gamepad1.left_stick_x, -gamepad1.right_stick_x/2, true);
+            follower.setTeleOpDrive(drive, strafe, turn, true);
             follower.update();
-
-            telemetry.addData("Turret Power:", turretPower);
-            telemetry.addData("Heading Error:", headingError);
-            telemetry.addData("Absolute Angle:", turretPos);
-
-            telemetry.update();
-            sleep(1);
+            sleep(10);
         }
     }
 
@@ -221,7 +235,7 @@ public class TurretAprilTagTracking extends LinearOpMode
         // Decimation = 3 ..  Detect 2" Tag from 4  feet away at 30 Frames Per Second
         // Decimation = 3 ..  Detect 5" Tag from 10 feet away at 30 Frames Per Second
         // Note: Decimation can be changed on-the-fly to adapt during a match.
-        aprilTag.setDecimation(3);
+        aprilTag.setDecimation(2);
 
         // Create the vision portal by using a builder.
         if (USE_WEBCAM) {
