--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/LOADCode/Testing_/TeleOps/TurretLocalization.java
@@ -1,357 +1,0 @@
-/* Copyright (c) 2023 FIRST. All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without modification,
- * are permitted (subject to the limitations in the disclaimer below) provided that
- * the following conditions are met:
- *
- * Redistributions of source code must retain the above copyright notice, this list
- * of conditions and the following disclaimer.
- *
- * Redistributions in binary form must reproduce the above copyright notice, this
- * list of conditions and the following disclaimer in the documentation and/or
- * other materials provided with the distribution.
- *
- * Neither the name of FIRST nor the names of its contributors may be used to endorse or
- * promote products derived from this software without specific prior written permission.
- *
- * NO EXPRESS OR IMPLIED LICENSES TO ANY PARTY'S PATENT RIGHTS ARE GRANTED BY THIS
- * LICENSE. THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
- * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE
- * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
- * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package org.firstinspires.ftc.teamcode.LOADCode.Testing_.TeleOps;
-
-import com.bylazar.configurables.annotations.Configurable;
-import com.bylazar.telemetry.TelemetryManager;
-import com.pedropathing.follower.Follower;
-import com.pedropathing.geometry.Pose;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import com.qualcomm.robotcore.hardware.AnalogInput;
-import com.qualcomm.robotcore.hardware.CRServo;
-
-import org.firstinspires.ftc.robotcore.external.hardware.camera.BuiltinCameraDirection;
-import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
-import org.firstinspires.ftc.robotcore.external.hardware.camera.controls.ExposureControl;
-import org.firstinspires.ftc.robotcore.external.hardware.camera.controls.GainControl;
-import org.firstinspires.ftc.teamcode.pedroPathing.Constants;
-import org.firstinspires.ftc.vision.VisionPortal;
-import org.firstinspires.ftc.vision.apriltag.AprilTagDetection;
-import org.firstinspires.ftc.vision.apriltag.AprilTagProcessor;
-
-import java.util.List;
-import java.util.concurrent.TimeUnit;
-
-import dev.nextftc.control.ControlSystem;
-import dev.nextftc.control.KineticState;
-import dev.nextftc.control.feedback.PIDCoefficients;
-
-/*
- * This OpMode illustrates using a camera to locate and drive towards a specific AprilTag.
- * The code assumes a Holonomic (Mecanum or X Drive) Robot.
- *
- * For an introduction to AprilTags, see the ftc-docs link below:
- * https://ftc-docs.firstinspires.org/en/latest/apriltag/vision_portal/apriltag_intro/apriltag-intro.html
- *
- * When an AprilTag in the TagLibrary is detected, the SDK provides location and orientation of the tag, relative to the camera.
- * This information is provided in the "ftcPose" member of the returned "detection", and is explained in the ftc-docs page linked below.
- * https://ftc-docs.firstinspires.org/apriltag-detection-values
- *
- * The drive goal is to rotate to keep the Tag centered in the camera, while strafing to be directly in front of the tag, and
- * driving towards the tag to achieve the desired distance.
- * To reduce any motion blur (which will interrupt the detection process) the Camera exposure is reduced to a very low value (5mS)
- * You can determine the best Exposure and Gain values by using the ConceptAprilTagOptimizeExposure OpMode in this Samples folder.
- *
- * The code assumes a Robot Configuration with motors named: front_left_drive and front_right_drive, back_left_drive and back_right_drive.
- * The motor directions must be set so a positive power goes forward on all wheels.
- * This sample assumes that the current game AprilTag Library (usually for the current season) is being loaded by default,
- * so you should choose to approach a valid tag ID.
- *
- * Under manual control, the left stick will move forward/back & left/right.  The right stick will rotate the robot.
- * Manually drive the robot until it displays Target data on the Driver Station.
- *
- * Press and hold the *Left Bumper* to enable the automatic "Drive to target" mode.
- * Release the Left Bumper to return to manual driving mode.
- *
- * Under "Drive To Target" mode, the robot has three goals:
- * 1) Turn the robot to always keep the Tag centered on the camera frame. (Use the Target Bearing to turn the robot.)
- * 2) Strafe the robot towards the centerline of the Tag, so it approaches directly in front  of the tag.  (Use the Target Yaw to strafe the robot)
- * 3) Drive towards the Tag to get to the desired distance.  (Use Tag Range to drive the robot forward/backward)
- *
- * Use DESIRED_DISTANCE to set how close you want the robot to get to the target.
- * Speed and Turn sensitivity can be adjusted using the SPEED_GAIN, STRAFE_GAIN and TURN_GAIN constants.
- *
- */
-@Disabled
-@Configurable
-@TeleOp(name="Turret Localization", group = "TestTeleOp")
-public class TurretLocalization extends LinearOpMode
-{
-    private static final boolean USE_WEBCAM = true;  // Set true to use a webcam, or false for a phone camera
-    int DESIRED_TAG_ID = 24;    // Choose the tag you want to approach or set to -1 for ANY tag.
-    private VisionPortal visionPortal;               // Used to manage the video source.
-    private AprilTagProcessor aprilTag;              // Used for managing the AprilTag detection process.
-    public static Follower follower;                 // Used for managing the PedroPathing path follower
-    public TelemetryManager telemetryM;              // Used for putting telemetry data on Panels
-
-    //Turret PID pidCoefficients
-    public static PIDCoefficients turretCoefficients = new PIDCoefficients(0.75, 0.005, 200);
-
-    public static int turretDeadZoneSize = 5;
-
-    // Contains the start Pose of our robot. This can be changed or saved from the autonomous period.
-    private final Pose startPose = new Pose(135.6,9.8, Math.toRadians(90));
-
-    @Override public void runOpMode()
-    {
-        boolean targetFound;        // Set to true when an AprilTag target is detected
-        double headingError = 0;    // The error in degrees in the turret's angle
-
-        // Used to store the hardware objects for the robot
-        CRServo turret;
-        AnalogInput turretEncoder;
-
-        double turretPos;           // Used to store the current angle of the turret
-        double turretPower;     // Used to store the calculated power to output to the turret servo
-        boolean runTurret = true;
-        boolean useTurretPID = false;
-        double targetAngle;
-        double lastTurretPos;
-        int wraparoundTrigger = 0;
-        int[] goalCoords = new int[2];
-
-        ControlSystem turretPID = ControlSystem.builder()
-                .posPid(new PIDCoefficients(turretCoefficients.kP/100000000, turretCoefficients.kI/100000000, turretCoefficients.kD/100000000))
-                .build();
-
-        initAprilTag();             // Initialize the Apriltag Detection process
-
-        if (USE_WEBCAM)
-            setManualExposure(6, 250);  // Use low exposure time to reduce motion blur
-
-        follower = Constants.createFollower(hardwareMap);   // Initializes the PedroPathing path follower
-        follower.setStartingPose(startPose);                // Sets the initial position of the robot on the field
-        follower.update();                                  // Applies the initialization
-
-        telemetry.addData("Camera preview on/off", "3 dots, Camera Stream");
-        telemetry.addData(">", "Touch START to start OpMode");
-        telemetry.update();
-
-        waitForStart(); // Wait for driver to press start
-
-        follower.startTeleopDrive();    // Activate the manual drive function for the drivetrain
-        follower.update();              // Apply the activation
-
-        // Fetch the actual hardware objects and store them in their respective variables
-        turret = hardwareMap.get(CRServo.class, "turret");
-        turretEncoder = hardwareMap.get(AnalogInput.class,"turretEncoder");
-        lastTurretPos = Math.abs(((turretEncoder.getVoltage() / 3.3) * 360) - 360);
-
-        while (opModeIsActive())
-        {
-
-            if (gamepad1.yWasPressed()){
-                if (DESIRED_TAG_ID == 24){
-                    DESIRED_TAG_ID = 20;
-                } else {
-                    DESIRED_TAG_ID = 24;
-                }
-            }
-
-            if (DESIRED_TAG_ID == 24){
-                goalCoords[0] = 144;
-                goalCoords[1] = 144;
-            }else if (DESIRED_TAG_ID == 20){
-                goalCoords[0] = 0;
-                goalCoords[1] = 144;
-            }
-
-            // Used to indicate whether or not a valid AprilTag has been detected
-            targetFound = false;
-            // Used to hold the data for a detected AprilTag
-            AprilTagDetection desiredTag = null;
-
-            // Read the Axon servo encoder position and convert it from a voltage to an angle in degrees
-            turretPos = Math.abs(((turretEncoder.getVoltage() / 3.3) * 360) - 360);
-
-            // Step through the list of detected tags and look for a matching tag
-            List<AprilTagDetection> currentDetections = aprilTag.getDetections();
-            for (AprilTagDetection detection : currentDetections) {
-                // Look to see if we have size info on this tag.
-                if (detection.metadata != null) {
-                    //  Check to see if we want to track towards this tag.
-                    if ((DESIRED_TAG_ID < 0) || (detection.id == DESIRED_TAG_ID)) {
-                        // Yes, we want to use this tag.
-                        targetFound = true;
-                        desiredTag = detection;
-                        // Determine heading error so we can use them to control the robot automatically.
-                        headingError = -desiredTag.ftcPose.bearing;
-                        break;  // don't look any further.
-                    } else {
-                        // This tag is in the library, but we do not want to track it right now.
-                        telemetry.addData("Skipping", "Tag ID %d is not desired", detection.id);
-                    }
-                } else {
-                    // This tag is NOT in the library, so we don't have enough information to track to it.
-                    telemetry.addData("Unknown", "Tag ID %d is not in TagLibrary", detection.id);
-                }
-            }
-
-            // Tell the driver what we see, and what to do.
-            if (DESIRED_TAG_ID == 24){
-                telemetry.addData("Goal Selected: ","Red");
-            } else {
-                telemetry.addData("Goal Selected: ","Blue");
-            }
-            if (targetFound) {
-                telemetry.addData("Found", "ID %d (%s)", desiredTag.id, desiredTag.metadata.name);
-                telemetry.addData("Range",  "%5.1f inches", desiredTag.ftcPose.range);
-                telemetry.addData("Bearing","%3.0f degrees", desiredTag.ftcPose.bearing);
-                telemetry.addData("Yaw","%3.0f degrees", desiredTag.ftcPose.yaw);
-            } else {
-                telemetry.addData("\n>","Drive using joysticks to find a valid target\n");
-            }
-
-            // This block finds the robot position from it's odometry data and calculates the angle between the selected goal and the robot, adjusting the camera to match that angle
-
-            double[] robotPose = {follower.getPose().getX(), follower.getPose().getY()};
-            targetAngle = Math.toDegrees(Math.atan2(goalCoords[1]-robotPose[1], goalCoords[0]-robotPose[0])) - Math.toDegrees(follower.getPose().getHeading());
-            targetAngle = (180 + targetAngle) % 360;
-            targetAngle = Math.min(Math.max(targetAngle, turretDeadZoneSize), 360-turretDeadZoneSize);
-            double angleDiff = (targetAngle - turretPos);
-            turretPower = -Math.pow(Math.min(Math.abs(angleDiff/60),1), (4f/3f)) * Math.signum(angleDiff);
-
-            if (gamepad1.aWasPressed()){
-                useTurretPID = !useTurretPID;
-            }
-            if (useTurretPID){
-                turretPID.setGoal(new KineticState(targetAngle));
-                turretPower = -turretPID.calculate(new KineticState(turretPos));
-                telemetry.addData("Turret Control System:", "PID");
-            }else{
-                telemetry.addData("Turret Control System:", "Non-PID");
-            }
-
-            if (gamepad1.dpad_left){
-                turretPower = 0.4;
-            }else if (gamepad1.dpad_right){
-                turretPower = -0.4;
-            }
-
-            if ((Math.abs(turretPos-lastTurretPos) >= 120) && (turretPos<350 || turretPos>10)){
-                wraparoundTrigger += (int)Math.signum(turretPos - 180);
-            }
-            if (wraparoundTrigger != 0) {
-                turretPower = Math.max(Math.min(-wraparoundTrigger, 0.2), -0.2);
-            }
-
-            if (gamepad1.bWasPressed()){
-                runTurret = !runTurret;
-            }
-            if(runTurret){
-                turret.setPower(turretPower);
-            }else{
-                turret.setPower(0);
-            }
-
-            // TODO We will work on Daniel's deadzone method next meeting, If that does not work, we will work on Ari's!
-
-
-            // Apply desired axes motions to the drivetrain
-            follower.setTeleOpDrive(-gamepad1.left_stick_y, -gamepad1.left_stick_x, -gamepad1.right_stick_x, true);
-            follower.update();
-
-            lastTurretPos = turretPos;
-
-            telemetry.addData("Target Angle", targetAngle);
-            telemetry.addData("Turret Angle", turretPos);
-            telemetry.addData("Turret Error", angleDiff);
-            telemetry.addData("Turret Power", turretPower);
-            telemetry.addData("Robot Heading", Math.toDegrees(follower.getPose().getHeading()));
-            telemetry.addData("Wraparound Trigger This Time: ", wraparoundTrigger);
-            telemetry.addData("-", "---------------------------------------------");
-            telemetry.addData("kP", turretCoefficients.kP);
-            telemetry.addData("kI", turretCoefficients.kI);
-            telemetry.addData("kD", turretCoefficients.kD);
-
-            telemetry.update();
-        }
-    }
-
-    /**
-     * Initialize the AprilTag processor.
-     */
-    private void initAprilTag() {
-        // Create the AprilTag processor by using a builder.
-        aprilTag = new AprilTagProcessor.Builder().build();
-
-        // Adjust Image Decimation to trade-off detection-range for detection-rate.
-        // e.g. Some typical detection data using a Logitech C920 WebCam
-        // Decimation = 1 ..  Detect 2" Tag from 10 feet away at 10 Frames per second
-        // Decimation = 2 ..  Detect 2" Tag from 6  feet away at 22 Frames per second
-        // Decimation = 3 ..  Detect 2" Tag from 4  feet away at 30 Frames Per Second
-        // Decimation = 3 ..  Detect 5" Tag from 10 feet away at 30 Frames Per Second
-        // Note: Decimation can be changed on-the-fly to adapt during a match.
-        aprilTag.setDecimation(3);
-
-        // Create the vision portal by using a builder.
-        if (USE_WEBCAM) {
-            visionPortal = new VisionPortal.Builder()
-                    .setCamera(hardwareMap.get(WebcamName.class, "Webcam 1"))
-                    .addProcessor(aprilTag)
-                    .build();
-        } else {
-            visionPortal = new VisionPortal.Builder()
-                    .setCamera(BuiltinCameraDirection.BACK)
-                    .addProcessor(aprilTag)
-                    .build();
-        }
-    }
-
-    /*
-     Manually set the camera gain and exposure.
-     This can only be called AFTER calling initAprilTag(), and only works for Webcams;
-    */
-    private void    setManualExposure(int exposureMS, int gain) {
-        // Wait for the camera to be open, then use the controls
-
-        if (visionPortal == null) {
-            return;
-        }
-
-        // Make sure camera is streaming before we try to set the exposure controls
-        if (visionPortal.getCameraState() != VisionPortal.CameraState.STREAMING) {
-            telemetry.addData("Camera", "Waiting");
-            telemetry.update();
-            while (!isStopRequested() && (visionPortal.getCameraState() != VisionPortal.CameraState.STREAMING)) {
-                sleep(20);
-            }
-            telemetry.addData("Camera", "Ready");
-            telemetry.update();
-        }
-
-        // Set camera controls unless we are stopping.
-        if (!isStopRequested())
-        {
-            ExposureControl exposureControl = visionPortal.getCameraControl(ExposureControl.class);
-            if (exposureControl.getMode() != ExposureControl.Mode.Manual) {
-                exposureControl.setMode(ExposureControl.Mode.Manual);
-                sleep(50);
-            }
-            exposureControl.setExposure((long)exposureMS, TimeUnit.MILLISECONDS);
-            sleep(20);
-            GainControl gainControl = visionPortal.getCameraControl(GainControl.class);
-            gainControl.setGain(gain);
-            sleep(20);
-        }
-    }
-}
